pyproject.toml
[project]
name = "heteromark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "gym>=0.26.2",
    "gymnasium>=1.2.2",
    "hydra-core>=1.3.2",
    "pettingzoo==1.22.4",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "smacv2",
    "tensordict>=0.10.0",
    "torchrl[utils]==0.10.0",
    "tqdm>=4.67.1",
]

# This tells setuptools to treat `src/` as the root
[tool.setuptools]
package-dir = {"" = "src"}

# Automatically find packages inside src/
[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
smacv2 = { git = "https://github.com/oxwhirl/smacv2.git" }

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "G"]
